[ADD] CodeCharts DialogView, PictureView, GridView in Fullscreen Mode, Scaling of Pictures to screen resolution, replacing of PictureView after given time
