+++ build.gradle.kts
@@ -16,7 +16,8 @@ version = "1.0-SNAPSHOT"
 repositories {
     mavenCentral()
     maven {
-        url = uri("https://plugins.gradle.org/m2/")
+        // url = uri("https://plugins.gradle.org/m2/")
+        url = uri("https://oss.sonatype.org/content/repositories/snapshots")
     }
 }
 
@@ -35,7 +36,7 @@ object Version {
     const val SLF4J = "1.7.30"
     const val KMONGO = "4.2.2"
     const val ASSERTK = "0.23"
-    const val TORNADOFX = "1.7.20"
+    const val TORNADOFX = "2.0.0-SNAPSHOT"
     const val SPOTLESS = "5.8.2"
     const val KTLINT = "9.4.1"
     const val TESTFX = "4.0.15-alpha"
