+++ src/main/code/de/tuchemnitz/se/exercise/core/graphics/Style.kt
@@ -1,11 +1,17 @@
 package de.tuchemnitz.se.exercise.core.graphics
 
+import de.tuchemnitz.se.exercise.core.graphics.MainApp
+import de.tuchemnitz.se.exercise.core.graphics.MainApp.Companion.IMAGE_PATH
 import javafx.geometry.Pos
+import javafx.scene.layout.BackgroundPosition
+import javafx.scene.layout.BackgroundRepeat
 import javafx.scene.paint.Color
 import tornadofx.Stylesheet
 import tornadofx.box
 import tornadofx.cssclass
+import tornadofx.multi
 import tornadofx.px
+import java.net.URI
 
 class Style : Stylesheet() {
 
@@ -17,6 +23,10 @@ class Style : Stylesheet() {
         val mainCenterStyle by cssclass()
         val mainBottomStyle by cssclass()
         val mainBottomButtonStyle by cssclass()
+        val ccGridWrapper by cssclass()
+        val ccPictureWrapper by cssclass()
+        val ccDialogWrapper by cssclass()
+        // const val IMAGE_PATH = "/penguin.png"
     }
 
     init {
@@ -56,6 +66,17 @@ class Style : Stylesheet() {
         s(mainWrapper) {
             prefHeight = 300.px
             prefWidth = 300.px
+        }
+
+        s(ccPictureWrapper) {
+            prefHeight = 1000.px
+            prefWidth = 1000.px
+            s(imageView)
+            alignment = Pos.CENTER
+        }
+
+        s(ccDialogWrapper) {
+
         }
     }
 }
