--- src/main/code/de/tuchemnitz/se/exercise/core/graphics/codecharts/EmptyView.kt
@@ -1,9 +1,0 @@
-package de.tuchemnitz.se.exercise.core.graphics.codecharts
-
-import tornadofx.View
-import tornadofx.borderpane
-
-class EmptyView : View("My View") {
-    override val root = borderpane {
-    }
-}
